Only use "find slime" if the slime directory exists.

+++ Makefile
@@ -359,8 +359,9 @@ dist2:
 
 clean:
 	rm -fr build *.dmg
-	find slime -name '*.fasl' -delete
-	find slime site-lisp -name '*.elc' -delete
+	test -d slime && find slime -name '*.fasl' -delete
+	test -d slime && find slime -name '*.elc' -delete
+	find site-lisp -name '*.elc' -delete
 
 scour: clean
 	rm -fr aquamacs sbcl slime systems site
